# Add Vector3::cos_to and related optimizations