--- a/core/math/transform_2d.cpp
+++ b/core/math/transform_2d.cpp
@@ -71,7 +71,7 @@ void Transform2D::rotate(const real_t p_angle) {
 
 real_t Transform2D::get_skew() const {
 	real_t det = basis_determinant();
-	return Math::acos(columns[0].normalized().dot(SIGN(det) * columns[1].normalized())) - (real_t)Math_PI * 0.5f;
+	return Math::acos(columns[0].cos_to(SIGN(det) * columns[1])) - (real_t)Math_PI * 0.5f;
 }
 
 void Transform2D::set_skew(const real_t p_angle) {