--- a/core/math/vector2.h
+++ b/core/math/vector2.h
@@ -95,6 +95,7 @@ struct _NO_DISCARD_ Vector2 {
 
 	real_t distance_to(const Vector2 &p_vector2) const;
 	real_t distance_squared_to(const Vector2 &p_vector2) const;
+	real_t cos_to(const Vector2 &p_vector2) const;
 	real_t angle_to(const Vector2 &p_vector2) const;
 	real_t angle_to_point(const Vector2 &p_vector2) const;
 	_FORCE_INLINE_ Vector2 direction_to(const Vector2 &p_to) const;