Adds a method to get the cosine of the angle between two vectors faster than `normalized().dot(b.normalized())`, and uses it where applicable.

Sample output from test project (intel i7-10750H):
~~~
a.normalized().dot(b.normalized()) :
  100 :
    9 usecs
  1000 :
    93 usecs
  10000 : 
    736 usecs
  100000 :
    7 msecs, 929 usecs
  1000000 :
    72 msecs, 735 usecs
  10000000 :
    724 msecs, 601 usecs
  100000000 :
    7 secs, 151 msecs, 508 usecs

a.cos_to(b):
  100 :
    4 usecs
  1000 :
    66 usecs
  10000 :
    328 usecs
  100000 :
    3 msecs, 263 usecs
  1000000 :
    31 msecs, 441 usecs
  10000000 :
    315 msecs, 199 usecs
  100000000 :
    2 secs, 976 msecs, 159 usecs
~~~

[test project](https://github.com/godotengine/godot/files/10474998/cos_to_normdot_comparison.zip)