@oenixphay42 2023-01-22
> Please check, if this function could also be used in:
> 
>     * Transform2D::get_skew()
> 
>     * Node3DEditorViewport::update_transform_gizmo_view()
> 
>     * Animation::_vector2_track_optimize_key(...) Line 3934
> 
> 
> You might also want to have a look at tests/scene/test_primitives.h if there are adjustments possible:
> 
>     * SUBCASE("Only 6 distinct normals.")
> 
>     * SUBCASE("[Primitive][Sphere] All normals lie in direction of corresponding point.")

Updated to use in those places.