@oenixphay42 2023-01-24
Fixed the missing tabs and added tests for the edge case.